requirements.txt
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
cssselect2==0.7.0
Django==4.2.1
django-ckeditor==6.5.1
django-filer==2.2.4
django-js-asset==2.0.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-ranged-response==0.2.0
django-render-partial==0.4
django-simple-captcha==0.5.17
easy-thumbnails==2.8.5
idna==3.4
isodate==0.6.1
lxml==4.9.2
Pillow==9.5.0
platformdirs==3.5.1
psycopg2==2.9.6
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
sweetify==2.3.1
tinycss2==1.2.1
tzdata==2023.3
Unidecode==1.1.2
urllib3==2.0.2
webencodings==0.5.1
zeep==4.2.1
